Update App.js
UPDATE: The reason why the github pages for this site was not working is because there was some sort of conflict between the dependencies. Since i have yarn installed globally, i assumed that there must have been some sort of conflict between yarn and firebase. So i downloaded a new create react app that does not contain yarn by typing this into the terminal: npx create-react-app ig --use-npm . This new react app was created with npm and it did not have the yarn lock inside it. Then i replace the src and public folders of this ig react app, with the src and public folders that are in this instagram depository. Then i followed the steps to deploy the app through firebase hosting, instead of through github pages, and the site was successfully deployed. You can find it here: https://instagram-a3c9d.web.app/ "